app/requirements.txt
flask
mysql-connector-python
kafka-python
pandas
gunicorn
pyhive
thrift
thrift-sasl
pure-sasl
